+++ src/Controller/api/UsersController.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Api;
 
+use App\Controller\AppController;
 use Cake\ORM\TableRegistry;
 
 class UsersController extends AppController
